renamed potion package to status effect
